Removed old versions and fixed file names
